requirements.txt
google-api-python-client
pymongo[srv]
python-dotenv
transformers[torch]
pandas
torch